refactor: move some stubs out
