+++ src/test/java/org/jd/core/v1/JavaOperatorPrecedenceTest.java
@@ -23,6 +23,7 @@ import org.jd.core.v1.service.fragmenter.javasyntaxtojavafragment.JavaSyntaxToJa
 import org.jd.core.v1.service.layouter.LayoutFragmentProcessor;
 import org.jd.core.v1.service.tokenizer.javafragmenttotoken.JavaFragmentToTokenProcessor;
 import org.jd.core.v1.service.writer.WriteTokenProcessor;
+import org.jd.core.v1.stub.NewOperatorPrecedence;
 import org.junit.Test;
 
 import java.util.Collections;
@@ -39,18 +40,11 @@ public class JavaOperatorPrecedenceTest extends TestCase {
 
     @Test
     public void testNewOperatorPrecedence() throws Exception {
-        class NewOperatorPrecedence {
-            void test() {}
-            public /* static */ void main(String ... args) {
-                new NewOperatorPrecedence().test();
-            }
-        }
-
         String internalClassName = NewOperatorPrecedence.class.getName().replace('.', '/');
         String source = decompile(new ClassPathLoader(), new PlainTextPrinter(), internalClassName);
 
         // Check decompiled source code
-        assertTrue(source.matches(PatternMaker.make(": 44 */", "new NewOperatorPrecedence().test();")));
+        assertTrue(source.matches(PatternMaker.make(": 8 */", "new NewOperatorPrecedence().test();")));
 
         // Recompile decompiled source code and check errors
         assertTrue(CompilerUtil.compile("1.8", new JavaSourceFileObject(internalClassName, source)));
