+++ src/test/java/org/jd/core/v1/stub/AutoboxingAndUnboxing.java
@@ -1,0 +1,8 @@
+package org.jd.core.v1.stub;
+
+public class  AutoboxingAndUnboxing {
+    void test() {
+        Integer intObj = 10;
+        int i = intObj;
+    }
+}
