+++ src/test/java/org/jd/core/v1/stub/NewOperatorPrecedence.java
@@ -1,0 +1,10 @@
+package org.jd.core.v1.stub;
+
+public class NewOperatorPrecedence {
+    void test() {
+    }
+
+    public /* static */ void main(String... args) {
+        new NewOperatorPrecedence().test();
+    }
+}
